Refactor TabListManager to re-enable Minecraft imports and adjust player state tracking. Temporarily disable event registration and update tasks due to import issues.
Enhance ExternalEconomyManager with improved error handling and integration management. Simplify integration initialization and configuration loading. Update balance methods to use BigDecimal for consistency.

Update KitManager to ensure economy transactions use BigDecimal for cost handling.

Revise ColorUtil to simplify color processing and remove legacy code handling. Implement basic color stripping functionality.

Modify MessageUtil to utilize TextColor for color tests, enhancing visual representation.

Add new shop-related messages to language files for better user feedback during shop creation and management.

Remove unused web dashboard and themes files.

Delete test server admin commands file as it is no longer needed.

Introduce AFKEventManager as a placeholder for future AFK event management, with plans to restore functionality when imports stabilize.

+++ src/main/java/com/zerog/neoessentials/managers/KitManager.java
@@ -13,6 +13,7 @@ import com.zerog.neoessentials.util.PermissionUtil;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
 
+import java.math.BigDecimal;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -106,7 +107,7 @@ public class KitManager {
         // Check cost
         if (kit.hasCost()) {
             EconomyManager economyManager = EconomyManager.getInstance();
-            if (!economyManager.hasBalance(player.getUUID(), kit.getCost())) {
+            if (!economyManager.hasBalance(player.getUUID(), BigDecimal.valueOf(kit.getCost()))) {
                 String message = LanguageManager.getInstance().getMessage(player, "neoessentials.kit.insufficient_funds", 
                     kitName, kit.getCost());
                 player.sendSystemMessage(net.minecraft.network.chat.Component.literal(message));
@@ -120,7 +121,7 @@ public class KitManager {
             // Deduct cost if applicable
             if (kit.hasCost()) {
                 EconomyManager economyManager = EconomyManager.getInstance();
-                economyManager.withdrawBalance(player.getUUID(), kit.getCost(), "Kit purchase: " + kitName);
+                economyManager.withdrawBalance(player.getUUID(), BigDecimal.valueOf(kit.getCost()), "Kit purchase: " + kitName);
             }
             
             // Execute kit commands if any
